refactore project route into separate file
