+++ src/routes/projects/parallel/IndexProjectParallel.ts
@@ -1,0 +1,10 @@
+import ProjectParallel from './ProjectParallel'
+import ProjectParallelCLogic from './clogic/ProjectParallelCLogic'
+import ProjectParallelConfig from './config/ProjectParallelConfig'
+import router from 'src/routes/auth'
+
+export default {
+    ProjectParallel,
+    ProjectParallelCLogic,
+    ProjectParallelConfig
+}
